forgot to float the prices

+++ bamazonSchema.sql
@@ -7,7 +7,7 @@ CREATE TABLE products(
   id INT NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(100) NOT NULL,
   department VARCHAR(45) NOT NULL,
-  price INT default 0,
+  price FLOAT default 0,
   stock_quantity INT default 0,
   PRIMARY KEY (id)   
 );
